rename folder components -> navbar, update import in layout.tsx
